# BUGFIX: Only set distinct on count clause if explicitely set to improve performance