@albe 2023-08-24
Or wait, I was never checking a `$query->count()` on implicit join results, but just the `count($result)`... https://github.com/neos/flow-development-collection/pull/1293/files#diff-b389146f7ff3bedefca76487b8b3421d40ebf42cb4f24e854d129897dacc03a8L269
So maaaaybeee this is indeed a bit more breaky than my last comment suggests. But breaky in the sense it behaves differently than before, though more logical, i.e. `->count()` would return 2 now, when there is a duplicate result/2 result rows.