@Sebobo 2023-08-25
Setting a query as distinct should always be explicitly set and not assumed in any case?
Doctrine already does more magic than wished for in some cases.