@Sebobo 2023-08-25
Just got results from my customers production system based only on this change:

<img width="479" alt="Bildschirmfoto 2023-08-25 um 09 04 16" src="https://github.com/neos/flow-development-collection/assets/596967/0d7b5cc4-5a3d-471d-a345-7c523a003aa9">